requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
aiosqlite>=0.20.0
httpx>=0.27.0
python-dotenv>=1.0.0
pydantic-settings>=2.0.0
hiero-sdk-python>=0.2.0
